refactor: pixel tracking
Reorganized pixels tracking, preping for a dedicated solution for native ios.

Essentially I copied to the same approach of push notifications, allowing us to isolate platform specific implementations.

I also got rid of the different remote scripts, I find them useless
