+++ packages/shared/src/contexts/PaymentContext.tsx
@@ -24,11 +24,11 @@ import { useAuthContext } from './AuthContext';
 import { plusSuccessUrl } from '../lib/constants';
 import { LogEvent } from '../lib/log';
 import { usePlusSubscription } from '../hooks';
-import { logPixelPayment } from '../lib/pixels';
 import { feature } from '../lib/featureManagement';
 import { PlusPriceType, PlusPriceTypeAppsId } from '../lib/featureValues';
 import { getPrice } from '../lib';
 import { useFeature } from '../components/GrowthBookProvider';
+import { PixelsContext } from './PixelsContext';
 
 export type ProductOption = {
   label: string;
@@ -77,6 +77,7 @@ export const PaymentContextProvider = ({
 }: PaymentContextProviderProps): ReactElement => {
   const router = useRouter();
   const { user, geo, isValidRegion: isPlusAvailable } = useAuthContext();
+  const { trackPayment } = useContext(PixelsContext);
   const planTypes = useFeature(feature.pricingIds);
   const [paddle, setPaddle] = useState<Paddle>();
   const { logSubscriptionEvent, isPlus } = usePlusSubscription();
@@ -123,7 +124,7 @@ export const PaymentContextProvider = ({
                 payment: event?.data.payment.method_details.type,
               },
             });
-            logPixelPayment(
+            trackPayment(
               event?.data.totals.total,
               event?.data.currency_code,
               event?.data?.transaction_id,
